+++ content/tweets/genre.mdx
@@ -4,7 +4,6 @@ handle: guidefari
 avatarUrl: https://res.cloudinary.com/hokaspokas/image/upload/v1661928720/bar_shpohc.png
 date: Sep 03 2022
 ---
-
 > Genre is a construct for commerce, not pleasure
 
-[Atlas](https://soundcloud.com/atlasglaas)
+[Atlas](https://soundcloud.com/atlasglaas) quoting [this book](https://www.amazon.com/Production-Critical-Listening-Engineering-Presents/dp/1138845949).
